Remove some code and migrate color-picker to inferno
