+++ src/components/text-input.tsx
@@ -49,8 +49,9 @@ const setPosition = (e?: HTMLInputElement, position?: number) => {
 }
 
 const setFocus = (e: HTMLInputElement, shouldFocus: boolean) => {
-  if (e && e !== document.activeElement && shouldFocus) e.focus()
-  if (!shouldFocus) e && e.blur()
+  (shouldFocus && e.focus()) || e.blur()
+  // if (e && e !== document.activeElement && shouldFocus) e.focus()
+  // if (!shouldFocus) e && e.blur()
 }
 
 const nopMaybe = (obj: object) =>
